refactor: 플레이어와 딜러의 공통 부모 구현
플레이어와 딜러에서 사용되는 동일한 로직의 중복을 제거
