+++ src/main/java/Dealer.java
@@ -1,0 +1,3 @@
+public class Dealer extends Participant {
+
+}
