+++ src/main/java/Player.java
@@ -1,40 +1,8 @@
-import java.util.ArrayList;
-import java.util.List;
-
-public class Player {
-    private final List<Card> cards = new ArrayList<>();
+public class Player extends Participant {
     private final PlayerName name;
 
     public Player(PlayerName name) {
         this.name = name;
-    }
-
-    public void receiveCard(Card card) {
-        cards.add(card);
-    }
-
-    public int calculateScore() {
-        int countOfAce = countAce();
-        int score = cards.stream()
-                .mapToInt(Card::getScore)
-                .sum();
-
-        return applyAce(score, countOfAce);
-    }
-
-    private int countAce() {
-        return (int) cards.stream()
-                .filter(Card::isAce)
-                .count();
-    }
-
-    private int applyAce(int score, int aceCount) {
-        while (score > 21 && aceCount > 0) {
-            score -= 10;
-            aceCount -= 1;
-        }
-
-        return score;
     }
 
 }
